Move context management into goose-providers crate
Compaction logic is now app-layer-free: config reads and fast-model
selection are injected via new CompactionModel trait and
CompactionSettings. Thin wrappers in goose keep existing call sites
unchanged. token_counter and usage_estimator moved along as
dependencies; compaction prompt templates are embedded in
goose-providers as the single source of truth.
