+++ crates/goose-providers/Cargo.toml
@@ -54,6 +54,11 @@ pkcs1 = { version = "0.7.5", default-features = false, features = ["pkcs8", "std
 pkcs8 = { version = "0.10", default-features = false, features = ["alloc", "std"], optional = true }
 sec1 = { version = "0.7", default-features = false, features = ["der", "pkcs8", "std"], optional = true }
 include_dir = { workspace = true }
+blake3 = { version = "1", default-features = false, features = ["std"] }
+indoc = { workspace = true }
+lru = { workspace = true }
+minijinja = { version = "2.18", default-features = false, features = ["loader", "multi_template", "serde"] }
+tiktoken-rs = { version = "0.12", default-features = false }
 
 [dev-dependencies]
 test-case = { workspace = true }
