+++ crates/goose-providers/src/context_mgmt/prompts/compaction_summary.md
