+++ crates/goose/src/prompt_template.rs
@@ -68,6 +68,29 @@ fn user_prompts_dir() -> PathBuf {
     Paths::config_dir().join("prompts")
 }
 
+/// Default content for templates whose canonical copy lives outside this
+/// crate (the compaction templates moved to `goose-providers`), falling back
+/// to the embedded prompts directory.
+fn default_template_content(name: &str) -> Option<String> {
+    match name {
+        "compaction.md" => Some(goose_providers::context_mgmt::COMPACTION_PROMPT_TEMPLATE.into()),
+        "compaction_summary.md" => {
+            Some(goose_providers::context_mgmt::COMPACTION_SUMMARY_TEMPLATE.into())
+        }
+        _ => CORE_PROMPTS_DIR
+            .get_file(name)
+            .map(|file| String::from_utf8_lossy(file.contents()).to_string()),
+    }
+}
+
+/// Returns the user's customized content for a registered template, if any.
+pub fn user_template_override(name: &str) -> Option<String> {
+    if !is_registered(name) {
+        return None;
+    }
+    std::fs::read_to_string(user_prompts_dir().join(name)).ok()
+}
+
 fn is_registered(name: &str) -> bool {
     TEMPLATE_REGISTRY.iter().any(|(n, _)| *n == name)
 }
@@ -121,13 +144,12 @@ pub fn render_template<T: Serialize>(name: &str, context: &T) -> Result<String, 
             )
         })?
     } else {
-        let file = CORE_PROMPTS_DIR.get_file(name).ok_or_else(|| {
+        default_template_content(name).ok_or_else(|| {
             MiniJinjaError::new(
                 minijinja::ErrorKind::TemplateNotFound,
                 format!("Built-in template '{}' not found", name),
             )
-        })?;
-        String::from_utf8_lossy(file.contents()).to_string()
+        })?
     };
 
     render_string(&template_str, context)
@@ -136,9 +158,7 @@ pub fn render_template<T: Serialize>(name: &str, context: &T) -> Result<String, 
 pub fn get_template(name: &str) -> Option<Template> {
     let (_, description) = TEMPLATE_REGISTRY.iter().find(|(n, _)| *n == name)?;
 
-    let default_content = CORE_PROMPTS_DIR
-        .get_file(name)
-        .map(|file| String::from_utf8_lossy(file.contents()).to_string())?;
+    let default_content = default_template_content(name)?;
 
     let user_path = user_prompts_dir().join(name);
     let user_content = if user_path.exists() {
@@ -193,9 +213,7 @@ pub fn list_templates() -> Vec<Template> {
     TEMPLATE_REGISTRY
         .iter()
         .filter_map(|(name, description)| {
-            let default_content = CORE_PROMPTS_DIR
-                .get_file(name)
-                .map(|file| String::from_utf8_lossy(file.contents()).to_string())?;
+            let default_content = default_template_content(name)?;
 
             let user_path = user_prompts_dir().join(name);
             let user_content = if user_path.exists() {
